requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
lxml==4.9.3
pandas==2.1.3
python-dateutil==2.8.2


